+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Shortcut recipes for Granite, Andesite, Diorite, Cobblestone, and Smooth Stone brick block family.
 - Smooth Stone Stairs
+- Undyed Terracotta Bricks
 
 ### Fixed
 - Soul Torch Levers now give off the correct amount of light.